Refactor and reorganised /band-placement/
